feat: advanced - 컴포넌트 분리 수정 및 Hook 작성
