+++ packages/assignment-7/src/App.tsx
@@ -1,27 +1,23 @@
-import { ChangeEvent, useEffect,  useState } from 'react';
-import {RepeatType, Event} from "./types/types.ts";
-import {getWeekDates} from "./utils/utilsFunc.ts";
-
+import { useEffect,  useState } from 'react';
+import {Event} from "./types/types.ts";
+import {getWeekDates, isOverlapping} from "./utils/utilsFunc.ts";
+import useEvents from "./hook/useEvents.ts";
+import useForm from "./hook/useForm.ts";
+import useNotice from "./hook/useNotice.ts";
 
 // Components
 import {
-  Alert,
-  AlertIcon,
-  AlertTitle,
   Box,
-  CloseButton,
   Flex,
   useInterval,
   useToast,
-  VStack,
 } from '@chakra-ui/react';
-import AlertConfirm from "./components/AlertConfirm.tsx";
-import EventForm from "./components/EventForm.tsx";
-import EventList from "./components/EventList";
+import AlertConfirm from "./components/alert/AlertConfirm.tsx";
+import AlertNotice from "./components/alert/AlertNotice.tsx";
+import EventForm from "./components/form/EventForm.tsx";
+import EventList from "./components/list/EventList.tsx";
 import Calendar from "./components/calendar/Calendar.tsx";
 
-
-const dummyEvents: Event[] = [];
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
@@ -51,56 +47,61 @@ const fetchHolidays = (year: number, month: number) => {
 
 
 function App() {
-  const [events, setEvents] = useState<Event[]>(dummyEvents);
-  const [title, setTitle] = useState('');
-  const [date, setDate] = useState('');
-  const [startTime, setStartTime] = useState('');
-  const [endTime, setEndTime] = useState('');
-  const [description, setDescription] = useState('');
-  const [location, setLocation] = useState('');
-  const [category, setCategory] = useState('');
   const [view, setView] = useState<'week' | 'month'>('month');
-  const [editingEvent, setEditingEvent] = useState<Event | null>(null);
-  const [isRepeating, setIsRepeating] = useState(false);
-  const [repeatType, setRepeatType] = useState<RepeatType>('none');
-  const [repeatInterval, setRepeatInterval] = useState(1);
-  const [repeatEndDate, setRepeatEndDate] = useState('');
-  const [notificationTime, setNotificationTime] = useState(10);
-  const [notifications, setNotifications] = useState<{ id: number; message: string }[]>([]);
-  const [notifiedEvents, setNotifiedEvents] = useState<number[]>([]);
-
+  const [currentDate, setCurrentDate] = useState(new Date());
+  const [searchTerm, setSearchTerm] = useState('');
+  const [holidays, setHolidays] = useState<{ [key: string]: string }>({});
   const [isOverlapDialogOpen, setIsOverlapDialogOpen] = useState(false);
   const [overlappingEvents, setOverlappingEvents] = useState<Event[]>([]);
 
+  const {events, fetchEvents, saveEvent, deleteEvent} = useEvents();
 
-  const [startTimeError, setStartTimeError] = useState<string | null>(null);
-  const [endTimeError, setEndTimeError] = useState<string | null>(null);
-
-  const [currentDate, setCurrentDate] = useState(new Date());
-
-  const [searchTerm, setSearchTerm] = useState('');
-  const [holidays, setHolidays] = useState<{ [key: string]: string }>({});
-
+  const {
+    title,
+    setTitle,
+    date,
+    setDate,
+    startTime,
+    endTime,
+    description,
+    setDescription,
+    location,
+    setLocation,
+    category,
+    setCategory,
+    editingEvent,
+    isRepeating,
+    setIsRepeating,
+    repeatType,
+    setRepeatType,
+    repeatInterval,
+    setRepeatInterval,
+    repeatEndDate,
+    setRepeatEndDate,
+    notificationTime,
+    setNotificationTime,
+    editEvent,
+    handleStartTimeChange,
+    handleEndTimeChange,
+    startTimeError,
+    endTimeError,
+    validateTime
+  } = useForm();
+  const {
+    notifications,
+    setNotifications,
+    notifiedEvents,
+    checkUpcomingEvents
+  } = useNotice(events)
   const toast = useToast();
 
-  const fetchEvents = async () => {
-    try {
-      const response = await fetch('/api/events');
-      if (!response.ok) {
-        throw new Error('Failed to fetch events');
-      }
-      const data = await response.json();
-      setEvents(data);
-    } catch (error) {
-      console.error('Error fetching events:', error);
-      toast({
-        title: "이벤트 로딩 실패",
-        status: "error",
-        duration: 3000,
-        isClosable: true,
-      });
-    }
+  // 겹치는 일정을 찾는 함수
+  const findOverlappingEvents = (newEvent: Event): Event[] => {
+    return events.filter(event =>
+        event.id !== newEvent.id && isOverlapping(event, newEvent)
+    );
   };
+
 
   const addOrUpdateEvent = async () => {
     if (!title || !date || !startTime || !endTime) {
@@ -150,181 +151,6 @@ function App() {
     }
   };
 
-  const saveEvent = async (eventData: Event) => {
-    try {
-      let response;
-      if (editingEvent) {
-        response = await fetch(`/api/events/${eventData.id}`, {
-          method: 'PUT',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(eventData),
-        });
-      } else {
-        response = await fetch('/api/events', {
-          method: 'POST',
-          headers: {
-            'Content-Type': 'application/json',
-          },
-          body: JSON.stringify(eventData),
-        });
-      }
-
-      if (!response.ok) {
-        throw new Error('Failed to save event');
-      }
-
-      await fetchEvents(); // 이벤트 목록 새로고침
-      setEditingEvent(null);
-      resetForm();
-      toast({
-        title: editingEvent ? "일정이 수정되었습니다." : "일정이 추가되었습니다.",
-        status: "success",
-        duration: 3000,
-        isClosable: true,
-      });
-    } catch (error) {
-      console.error('Error saving event:', error);
-      toast({
-        title: "일정 저장 실패",
-        status: "error",
-        duration: 3000,
-        isClosable: true,
-      });
-    }
-  };
-
-  const deleteEvent = async (id: number) => {
-    try {
-      const response = await fetch(`/api/events/${id}`, {
-        method: 'DELETE',
-      });
-
-      if (!response.ok) {
-        throw new Error('Failed to delete event');
-      }
-
-      await fetchEvents(); // 이벤트 목록 새로고침
-      toast({
-        title: "일정이 삭제되었습니다.",
-        status: "info",
-        duration: 3000,
-        isClosable: true,
-      });
-    } catch (error) {
-      console.error('Error deleting event:', error);
-      toast({
-        title: "일정 삭제 실패",
-        status: "error",
-        duration: 3000,
-        isClosable: true,
-      });
-    }
-  };
-
-  const checkUpcomingEvents = async () => {
-    const now = new Date();
-    const upcomingEvents = events.filter(event => {
-      const eventStart = new Date(`${event.date}T${event.startTime}`);
-      const timeDiff = (eventStart.getTime() - now.getTime()) / (1000 * 60);
-      return timeDiff > 0 && timeDiff <= event.notificationTime && !notifiedEvents.includes(event.id);
-    });
-
-    for (const event of upcomingEvents) {
-      try {
-        setNotifications(prev => [...prev, {
-          id: event.id,
-          message: `${event.notificationTime}분 후 ${event.title} 일정이 시작됩니다.`
-        }]);
-        setNotifiedEvents(prev => [...prev, event.id]);
-      } catch (error) {
-        console.error('Error updating notification status:', error);
-      }
-    }
-  };
-
-  const validateTime = (start: string, end: string) => {
-    if (!start || !end) return;
-
-    const startDate = new Date(`2000-01-01T${start}`);
-    const endDate = new Date(`2000-01-01T${end}`);
-
-    if (startDate >= endDate) {
-      setStartTimeError("시작 시간은 종료 시간보다 빨라야 합니다.");
-      setEndTimeError("종료 시간은 시작 시간보다 늦어야 합니다.");
-    } else {
-      setStartTimeError(null);
-      setEndTimeError(null);
-    }
-  };
-
-  const handleStartTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newStartTime = e.target.value;
-    setStartTime(newStartTime);
-    validateTime(newStartTime, endTime);
-  };
-
-  const handleEndTimeChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const newEndTime = e.target.value;
-    setEndTime(newEndTime);
-    validateTime(startTime, newEndTime);
-  };
-
-
-  // 날짜 문자열을 Date 객체로 변환하는 함수
-  const parseDateTime = (date: string, time: string): Date => {
-    return new Date(`${date}T${time}`);
-  };
-
-  // 두 일정이 겹치는지 확인하는 함수
-  const isOverlapping = (event1: Event, event2: Event): boolean => {
-    const start1 = parseDateTime(event1.date, event1.startTime);
-    const end1 = parseDateTime(event1.date, event1.endTime);
-    const start2 = parseDateTime(event2.date, event2.startTime);
-    const end2 = parseDateTime(event2.date, event2.endTime);
-
-    return start1 < end2 && start2 < end1;
-  };
-
-  // 겹치는 일정을 찾는 함수
-  const findOverlappingEvents = (newEvent: Event): Event[] => {
-    return events.filter(event =>
-        event.id !== newEvent.id && isOverlapping(event, newEvent)
-    );
-  };
-
-  const resetForm = () => {
-    setTitle('');
-    setDate('');
-    setStartTime('');
-    setEndTime('');
-    setDescription('');
-    setLocation('');
-    setCategory('');
-    setEditingEvent(null);
-    setIsRepeating(false);
-    setRepeatType('none');
-    setRepeatInterval(1);
-    setRepeatEndDate('');
-  };
-
-  const editEvent = (event: Event) => {
-    setEditingEvent(event);
-    setTitle(event.title);
-    setDate(event.date);
-    setStartTime(event.startTime);
-    setEndTime(event.endTime);
-    setDescription(event.description);
-    setLocation(event.location);
-    setCategory(event.category);
-    setIsRepeating(event.repeat.type !== 'none');
-    setRepeatType(event.repeat.type);
-    setRepeatInterval(event.repeat.interval);
-    setRepeatEndDate(event.repeat.endDate || '');
-    setNotificationTime(event.notificationTime);
-  };
-
   const navigate = (direction: 'prev' | 'next') => {
     setCurrentDate(prevDate => {
       const newDate = new Date(prevDate);
@@ -361,6 +187,10 @@ function App() {
       return true;
     })
   })();
+
+
+
+
 
   useInterval(checkUpcomingEvents, 1000); // 1초마다 체크
 
@@ -452,18 +282,7 @@ function App() {
             notificationTime={notificationTime}
         />
 
-        {notifications.length > 0 && <VStack position="fixed" top={4} right={4} spacing={2} align="flex-end">
-          {notifications.map((notification, index) => (
-              <Alert key={index} status="info" variant="solid" width="auto" data-testid="alert">
-                <AlertIcon/>
-                <Box flex="1">
-                  <AlertTitle fontSize="sm">{notification.message}</AlertTitle>
-                </Box>
-                <CloseButton onClick={() => setNotifications(prev => prev.filter((_, i) => i !== index))}/>
-              </Alert>
-          ))}
-        </VStack>
-        }
+        {notifications.length > 0 && <AlertNotice notifications={notifications} setNotifications={setNotifications} />}
       </Box>
   );
 }
